Name docker image tag more specifically
I think the issue we've had with deployment are a result of two builds running simultaneously, tagging the images with the same name, on the same DOCKER_HOST. The `asl` project ended up deploying an `asl-public-api` image.

Renaming the tagged image to match the name of the project should mitigate against this.