Fix bug with name search removing role filters
Fix was in @asl/schema, so update to include new version.

Improve unit tests by adding a new nacwo with the same name as the PEL holder. This ensures that both the name filter and role filter need to have been correctly applied to receive the expected result.

Fix other tests to reflect the addition of a new nacwo, and make some assertions stronger.