Fix deployment config for rate limiter
Environment variables are _always_ strings, so we need to do a little quote dance to make sure that the number is interpreted as a string when parsing into deploy config (TODO: fix necessity for that elsewhere).

Additionally ensure that string environment variables are parsed back into ints before trying to use them.