Add profle edit endpoint not scoped under establishment
A user should be able to edit _their own_ profile without having to be inside the context of a particular establishment.

This adds a PUT method to `/me` which will save to the profile of the _current_ logged in user, rather than a specfic establishment user fully specified in the url.