Refactor workflow client to take single set of parameters
Rather than passing some parameters on a request object, and others separately, the workflow client accepts a single set of parameters for performing an action.

Scoped task-related actions under a `.task(taskId)` function to help make it clearer when actions are performed on licence objects and when they are performed on tasks.