Fix establishment scoping of permissions removal
`req.establishmentId` is not set, which prevents the profile being scoped to an establishemnt and so rejects an attempt to remove permissions from a person with roles at other etablishments.

By correctly setting the establishemnt id in the scope call the check for removal of permissions is successful and a person's permissions can be removed.