Set RA properties on project API responses
Instead of doing the calculations of RA based on project state in the UI at render time, do the logic here and reduce the need for sending large payloads of data and js to the client.

Logic for RA has been copied directly from asl/projects. Once this is in place then it can be removed from there and make use of the API properties instead.