Make id parameter in getOpenTasks middleware immutable
Mutating the `id` by assigning the result of a function call to it would overwrite the id parameter for subsequent calls to the same function, and so once one project version had been loaded, that same id would be used for all subsequent calls to the middleware.

Use a separate internal variable to store the id within the scope of a single request, to avoid mutating the router level parameter.