Fix billing for users with multiple training PILs
If a user has held two training PILs within the same year at different establishments then when querying the list of billable PILs is it possible that there are training PIL records returned with the profile where the course was at a different establishment, and so the training course details are not attached to the record.

Check that the training course record is available before reading the establishment id.