Don't wait for profile last login touch
Currently if the request to update the last activity timestamp fails then the entire API service crashes because there's no error handler.

Instead swallow the error completely and get a perf boost by not waiting for the response from the request to update the timestamp before continuing.

Because the profile is refetched on every POST this should provide a reasonable performance boost on journeys which do frequent POST requests (a.k.a. ROPs).